Began making project linux compatible Unified file names to lower-case, added/replaced windows with linux header includes, Began converting specialized nested function templates to partially specialized(adding Dummy)

+++ src/font.h
@@ -1,6 +1,8 @@
 #pragma once
-#include <glm\glm.hpp>
+#include <glm.hpp>
 #include <vector>
+#include <string>
+
 namespace gui {
 	namespace text {
 
@@ -13,7 +15,7 @@ namespace gui {
 
 		struct Font {
 			Font() :instructions(0), kerningOffset(0), kerningCount(0), metricOffset(0), metricCount(0), glyphStorageIndex(0), atlasID(0), fontMetric(FontMetric()), stringOffset(0), stringCount(0) {}
-			size_t atlasID;
+			unsigned int atlasID;
 			FontMetric fontMetric;
 			size_t instructions;
 			size_t glyphStorageIndex;
@@ -89,8 +91,8 @@ namespace gui {
 			glm::vec2 pos;
 			glm::vec2 size;
 		};
-		extern size_t fontVAO;
-		extern size_t glyphShaderProgram;
+		extern unsigned int fontVAO;
+		extern unsigned int glyphShaderProgram;
 		extern std::vector<size_t> glyphIndexBuffer;
 		extern std::vector<Font> allFonts;
